+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_11_15_155736) do
+ActiveRecord::Schema.define(version: 2020_11_15_160323) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -30,6 +30,15 @@ ActiveRecord::Schema.define(version: 2020_11_15_155736) do
     t.jsonb "name", default: {}, null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+  end
+
+  create_table "foundations", force: :cascade do |t|
+    t.string "ticker", limit: 255
+    t.jsonb "name", default: {}, null: false
+    t.string "uuid", default: "gen_random_uuid()", null: false
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.index ["ticker"], name: "index_foundations_on_ticker", unique: true
   end
 
   create_table "shares", force: :cascade do |t|