+++ spec/factories/foundations.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :foundation do
+    sequence(:ticker) { |i| "YNDX#{i}" }
+    name { { en: 'Yandex Foundation', ru: 'Яндекс Фонд' } }
+  end
+end